+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.6] - 2019-06-27
+
+### Added
+
+- Added whois lookups for creation and modification dates 
+
+### Changed
+
+- Updated GeoLite database
+
 ## [1.2.5] - 2019-05-04
 
 ### Added